Major optimizations for fps, input, and performance
Previously tried to fix input delay for the menu and it worked; however,
it still left the game running poorly and the tile editor was slow.

Managed to fix these issues and keep the frames per second at 60.

Also, found that it'd be more optimal to convert all surfaces (images)
before blitting them to the screen.

Added another column of button choices for the sidebar in
tile_helper.py.